Improve Update Resilience
Currently when --update is used, changes are fetched from
the "origin" remote. This does not take into account changes
in the upstream repo url.

This commit ignores the remote, and fetches directly from the
url supplied. This avoids maintaining the remote URLs while always
using the correct upstream repo.

This commit also avoids using pull, as that requires a clean merge
and may fail if the repo is dirty or in some unexpected state. Instead
fetch and reset is used, as that will work in (almost) any repo state.

Additional clean and checkout commands are added to ensure the repo
is tidy.

Checkout is switched from the gitpython wrapper to a direct invocation
as the wrapper didn't auto create a local tracking branch when checking
out a remote branch.

Updates fetch test to monkey patch the git command executor.