RE-622 Fix GitHub URLs ending with .git in report
This change strips `".git"` from GitHub URLs used in the report.

The report produced by osa-differ/rpc-differ includes links to the
individual commits on GitHub. If a URL supplied on the command-line or
defined in an Ansible role file ended with `".git"`, the links generated
for the report would fail with a 404 without this change.

Without this change:

Valid commit links only generated from a URL not ending in `".git"`:
https://github.com/major/osa_differ/commit/SHA

Invalid commit links generated from a URL ending in `".git"`:
https://github.com/major/osa_differ.git/commit/SHA