First validate that role repositories exist
Recently ceph-ansible deleted roles from github,
causing osa-differ to fail when trying to check
for changes between two refs.

To cater to this situation, we add a check to see
whether a role repository exists. If not, it does
not include it in the list of roles to compare.

JIRA: RI-641